Removing warning suppression flags from pc/.
Bug: webrtc:9251

+++ pc/datachannel_unittest.cc
@@ -172,7 +172,8 @@ TEST_F(SctpDataChannelTest, BufferedAmountWhenBlocked) {
   }
   EXPECT_EQ(buffer.data.size() * number_of_packets,
             webrtc_data_channel_->buffered_amount());
-  EXPECT_EQ(number_of_packets, observer_->on_buffered_amount_change_count());
+  EXPECT_EQ(static_cast<size_t>(number_of_packets),
+            observer_->on_buffered_amount_change_count());
 }
 
 // Tests that the queued data are sent when the channel transitions from blocked
